+++ packages/ndk/lib/src/web_socket_client/src/_web_socket_connect/_web_socket_connect.dart
@@ -1,0 +1,11 @@
+/// Create a WebSocket connection.
+Future<Stream<dynamic>> connect(
+  String url, {
+  Iterable<String>? protocols,
+  Map<String, dynamic>? headers,
+  Duration? pingInterval,
+  String? binaryType,
+  bool compressionEnabled = true,
+}) {
+  throw UnsupportedError('No implementation of the api provided');
+}